.addProduct() Improved
'.addProduct()' now maintains a single object within the shopping cart for each product instead of multiple instances of the same product.

'productQuantity()' finds a product in the shopping cart and returns its quantity value +1.

The original shopping cart is filtered of the product being added, and the new product with an increased quantity replaces the old one in the array before being stored in localstorage.